requirements.txt
autopep8==1.5.7
cramjam==2.4.0
fastparquet==0.7.1
fsspec==2021.9.0
grpcio==1.41.0
grpcio-tools==1.41.0
numpy==1.21.2
pandas==1.3.3
protobuf==3.18.0
pycodestyle==2.7.0
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
thrift==0.15.0
toml==0.10.2
